Fix compiler warnings for gcc47
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@155609 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 6ca206d51561daed09c3c3d6621c12c11af1a7cc

+++ Online/Online/RootCnv/src/RootEvtSelector.cpp
@@ -277,7 +277,7 @@ RootEvtSelector::createAddress(const Context& ctxt, IOpaqueAddress*& pAddr) cons
       RootEvtSelectorContext::Files::const_iterator fileit = pctxt->fileIterator();
       if ( fileit != pctxt->files().end() ) {
         const string par[2] = {pctxt->fid(), m_rootName};
-        const unsigned long ipar[2] = {0,ent};
+        const unsigned long ipar[2] = {0,(unsigned long)ent};
         return m_dbMgr->createAddress(m_dbMgr->repSvcType(),m_rootCLID,&par[0],&ipar[0],pAddr);
       }
     }
